backend/requirements.txt
--pre
--extra-index-url https://download.pytorch.org/whl/nightly/cu128   
torch
torchaudio
torchvision

numpy<2.0
pyannote.audio==3.1.1
openai-whisper @ git+https://github.com/openai/whisper.git
celery[redis]==5.4.0
tqdm
openai
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart
